requirements.txt
asgiref==3.10.0
Django==4.1.4
django-ckeditor==6.5.1
django-js-asset==3.1.2
pillow==12.0.0
sqlparse==0.5.3
tzdata==2025.2
